Refactor code and remove warnings
