+++ sea-orm-macros/src/derives/simple_input.rs
@@ -1,18 +1,10 @@
 use bae::FromAttributes;
-use heck::CamelCase;
 use proc_macro2::{Ident, TokenStream};
 use quote::{format_ident, quote, quote_spanned};
-use syn::{
-    punctuated::Punctuated, token::Comma, Data, DataStruct, DeriveInput, Field, Fields,
-    FieldsNamed, Generics, Result,
-};
-
-use crate::{
-    derives::simple_input::field_validation::expand_field_validation,
-    util::option_type_to_inner_type,
-};
+use syn::{Data, DataStruct, DeriveInput, Fields, FieldsNamed, Result};
 
 mod field_validation;
+mod insertable;
 
 #[derive(FromAttributes)]
 struct Input {
@@ -40,62 +32,11 @@ pub(crate) fn expand_derive_simple_input(input: DeriveInput) -> Result<TokenStre
         }
     };
 
-    let insertable = impl_insertable(&ident, input.generics, &entity_ident, &fields)?;
-    let field_validation = expand_field_validation(&model_ident, &fields)?;
+    let insertable = insertable::impl_insertable(&ident, input.generics, &entity_ident, &fields);
+    let field_validation = field_validation::expand_field_validation(&model_ident, &fields);
 
-    let expanded = quote!(
+    Ok(quote!(
         #insertable
-
         #field_validation
-    );
-
-    Ok(expanded)
-}
-
-fn impl_insertable(
-    input_model_ident: &Ident,
-    mut input_model_generics: Generics,
-    entity_ident: &Ident,
-    fields: &Punctuated<Field, Comma>,
-) -> Result<TokenStream> {
-    input_model_generics
-        .lifetimes_mut()
-        .into_iter()
-        .for_each(|mut lifetime| {
-            lifetime.lifetime.ident = format_ident!("_");
-        });
-
-    let get_fields = fields.iter().map(|field| {
-        let field_name = field.ident.as_ref().unwrap();
-        let column_name = format_ident!("{}", field_name.to_string().to_camel_case());
-
-        if option_type_to_inner_type(&field.ty).is_some() {
-            quote!(
-                <Self::Entity as sea_orm::entity::EntityTrait>::Column::#column_name => {
-                    if let Some(value) = &self.#field_name {
-                        sea_orm::ActiveValue::set(value.clone()).into_wrapped_value()
-                    } else {
-                        sea_orm::ActiveValue::unset()
-                    }
-                }
-            )
-        } else {
-            quote!(<Self::Entity as sea_orm::entity::EntityTrait>::Column::#column_name => sea_orm::ActiveValue::set(self.#field_name.clone()).into_wrapped_value())
-        }
-    });
-
-    let expanded = quote!(
-        impl sea_orm::Insertable for #input_model_ident#input_model_generics {
-            type Entity = #entity_ident;
-
-            fn take(&mut self, c: <Self::Entity as sea_orm::entity::EntityTrait>::Column) -> sea_orm::ActiveValue<sea_orm::Value> {
-                match c {
-                    #(#get_fields,)*
-                    _ => sea_orm::ActiveValue::unset(),
-                }
-            }
-        }
-    );
-
-    Ok(expanded)
+    ))
 }
