+++ src/executor/query.rs
@@ -1,4 +1,6 @@
-use crate::{debug_print, DbErr};
+#[cfg(feature = "mock")]
+use crate::debug_print;
+use crate::DbErr;
 use std::fmt;
 
 #[derive(Debug)]
@@ -122,7 +124,9 @@ macro_rules! try_getable_all {
 macro_rules! try_getable_unsigned {
     ( $type: ty ) => {
         impl TryGetable for $type {
+            #[allow(unused_variables)]
             fn try_get(res: &QueryResult, pre: &str, col: &str) -> Result<Self, TryGetError> {
+                #[cfg(not(feature = "sqlx-postgres"))]
                 let column = format!("{}{}", pre, col);
                 match &res.row {
                     #[cfg(feature = "sqlx-mysql")]
@@ -157,7 +161,9 @@ macro_rules! try_getable_unsigned {
 macro_rules! try_getable_mysql {
     ( $type: ty ) => {
         impl TryGetable for $type {
+            #[allow(unused_variables)]
             fn try_get(res: &QueryResult, pre: &str, col: &str) -> Result<Self, TryGetError> {
+                #[cfg(not(feature = "sqlx-postgres"))]
                 let column = format!("{}{}", pre, col);
                 match &res.row {
                     #[cfg(feature = "sqlx-mysql")]
@@ -190,6 +196,7 @@ macro_rules! try_getable_postgres {
     ( $type: ty ) => {
         impl TryGetable for $type {
             fn try_get(res: &QueryResult, pre: &str, col: &str) -> Result<Self, TryGetError> {
+                #[cfg(not(feature = "sqlx-sqlite"))]
                 let column = format!("{}{}", pre, col);
                 match &res.row {
                     #[cfg(feature = "sqlx-mysql")]
